switch to double accuracy cost thresholds

+++ mcts/cost_constrained/cost_constrained_statistic.h
@@ -465,7 +465,7 @@ private:
 
     const std::vector<bool> use_cost_thresholding_;
     const std::vector<bool> use_chance_constrained_updates_;
-    const std::vector<float> cost_tresholds_;
+    const std::vector<double> cost_tresholds_;
     const bool use_lambda_policy_;
 };
